Added implementation for images
